Basic UI, Terrain Generation, Resources
Added three new resource tiles into the game: iron, gold, and wheat/grain. Improved the terrain generation so that different tiles generate with different frequencies (iron is much rarer than trees, for example). Also started on some basic UI which can display information about the game, including the status of a mobile unit and the movement cost of a tile.